Rename 'weight' to 'importance'

+++ NotesOrganiser-WebClient/src/app/components/note/edit-note/edit-note.component.ts
@@ -58,17 +58,19 @@ export class EditNoteComponent implements OnInit {
   }
 
   deleteData(id) {
-    this.noteService.deleteData(id).subscribe(
-      res => {
-        this.data = res;
-        this.toastr.error(JSON.stringify(this.data.message), "", {
-          timeOut: 2000,
-          progressBar: true,
-          positionClass: "toast-bottom-right"
-        });
+    if(confirm("Are you sure that you want to delete this note?")) {
+      this.noteService.deleteData(id).subscribe(
+        res => {
+          this.data = res;
+          this.toastr.error(JSON.stringify(this.data.message), "", {
+            timeOut: 2000,
+            progressBar: true,
+            positionClass: "toast-bottom-right"
+          });
 
-        this._router.navigateByUrl('/notes');
-      }
-    );
+          this._router.navigateByUrl('/notes');
+        }
+      );
+    }
   }
 }
